Update sklearn_extensions to latest

+++ select_model.py
@@ -1636,7 +1636,7 @@ parser.add_argument('--rna-trf-ft', type=str, nargs='+',
                     help='RNA trf fit type')
 parser.add_argument('--rna-trf-mb', type=str_bool, nargs='+',
                     help='RNA trf model batch')
-parser.add_argument('--rna-trf-pc', type=int, nargs='+',
+parser.add_argument('--rna-trf-pc', type=float, nargs='+',
                     help='RNA trf prior count')
 parser.add_argument('--nsn-trf-cc', type=str, nargs='+',
                     help='NanoStringNormalizer code_count')
